+++ Watchman.Discord/Areas/Protection/Services/AntiSpamService.cs
@@ -5,8 +5,8 @@ using System;
 using System.Threading.Tasks;
 using Devscord.DiscordFramework.Services.Factories;
 using Watchman.Common.Models;
-using Watchman.DomainModel.Protection;
 using Devscord.DiscordFramework.Framework.Commands.AntiSpam.Models;
+using Watchman.DomainModel.Protection.Mutes;
 
 namespace Watchman.Discord.Areas.Protection.Services
 {
