+++ Watchman.Discord/Areas/Protection/Services/PunishmentsCachingService.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Threading.Tasks;
 using Devscord.DiscordFramework.Framework.Commands.AntiSpam.Models;
 using Watchman.Cqrs;
-using Watchman.DomainModel.Protection;
-using Watchman.DomainModel.Protection.Commands;
-using Watchman.DomainModel.Protection.Queries;
+using Watchman.DomainModel.Protection.Antispam;
+using Watchman.DomainModel.Protection.Antispam.Commands;
+using Watchman.DomainModel.Protection.Antispam.Queries;
 
 namespace Watchman.Discord.Areas.Protection.Services
 {
