+++ Watchman.DomainModel/Protection/Antispam/Commands/Handlers/AddProtectionPunishmentCommandHandler.cs
@@ -1,0 +1,22 @@
+﻿using System.Threading.Tasks;
+using Watchman.Cqrs;
+using Watchman.Integrations.MongoDB;
+
+namespace Watchman.DomainModel.Protection.Antispam.Commands.Handlers
+{
+    public class AddProtectionPunishmentCommandHandler : ICommandHandler<AddProtectionPunishmentCommand>
+    {
+        private readonly ISessionFactory _sessionFactory;
+
+        public AddProtectionPunishmentCommandHandler(ISessionFactory sessionFactory)
+        {
+            this._sessionFactory = sessionFactory;
+        }
+
+        public async Task HandleAsync(AddProtectionPunishmentCommand command)
+        {
+            using var session = this._sessionFactory.Create();
+            await session.AddAsync(command.ProtectionPunishment);
+        }
+    }
+}
