+++ Watchman.DomainModel/Protection/Antispam/ProtectionPunishmentOption.cs
@@ -1,0 +1,9 @@
+﻿namespace Watchman.DomainModel.Protection.Antispam
+{
+    public enum ProtectionPunishmentOption
+    {
+        Nothing = 1,
+        Warn = 2,
+        Mute = 4
+    }
+}
