+++ Watchman.DomainModel/Protection/Antispam/Queries/GetProtectionPunishmentsQuery.cs
@@ -1,0 +1,8 @@
+﻿using Watchman.Cqrs;
+
+namespace Watchman.DomainModel.Protection.Antispam.Queries
+{
+    public class GetProtectionPunishmentsQuery : IQuery<GetProtectionPunishmentsQueryResult>
+    {
+    }
+}
