--- Watchman.DomainModel/Protection/Commands/AddMuteEventCommand.cs
@@ -1,14 +1,0 @@
-﻿using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Protection.Commands
-{
-    public class AddMuteEventCommand : ICommand
-    {
-        public MuteEvent MuteEvent { get; }
-
-        public AddMuteEventCommand(MuteEvent muteEvent)
-        {
-            this.MuteEvent = muteEvent;
-        }
-    }
-}
