--- Watchman.DomainModel/Protection/Commands/MarkMuteEventAsUnmutedCommand.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Protection.Commands
-{
-    public class MarkMuteEventAsUnmutedCommand : ICommand
-    {
-        public Guid MuteEventGuid { get; }
-
-        public MarkMuteEventAsUnmutedCommand(Guid muteEventGuid)
-        {
-            this.MuteEventGuid = muteEventGuid;
-        }
-    }
-}
