+++ Watchman.DomainModel/Protection/Mutes/Queries/GetMuteEventsQueryResult.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+using Watchman.Cqrs;
+
+namespace Watchman.DomainModel.Protection.Mutes.Queries
+{
+    public class GetMuteEventsQueryResult : IQueryResult
+    {
+        public IEnumerable<MuteEvent> MuteEvents { get; }
+
+        public GetMuteEventsQueryResult(IEnumerable<MuteEvent> muteEvents)
+        {
+            this.MuteEvents = muteEvents;
+        }
+    }
+}
