--- Watchman.DomainModel/Protection/ProtectionPunishmentOption.cs
@@ -1,9 +1,0 @@
-﻿namespace Watchman.DomainModel.Protection
-{
-    public enum ProtectionPunishmentOption
-    {
-        Nothing = 1,
-        Warn = 2,
-        Mute = 4
-    }
-}
