--- Watchman.DomainModel/Protection/Queries/GetProtectionPunishmentsQuery.cs
@@ -1,8 +1,0 @@
-﻿using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Protection.Queries
-{
-    public class GetProtectionPunishmentsQuery : IQuery<GetProtectionPunishmentsQueryResult>
-    {
-    }
-}
