--- Watchman.DomainModel/Protection/Queries/GetProtectionPunishmentsQueryResult.cs
@@ -1,15 +1,0 @@
-﻿using System.Collections.Generic;
-using Watchman.Cqrs;
-
-namespace Watchman.DomainModel.Protection.Queries
-{
-    public class GetProtectionPunishmentsQueryResult : IQueryResult
-    {
-        public IEnumerable<ProtectionPunishment> ProtectionPunishments { get; }
-
-        public GetProtectionPunishmentsQueryResult(IEnumerable<ProtectionPunishment> protectionPunishments)
-        {
-            this.ProtectionPunishments = protectionPunishments;
-        }
-    }
-}
